% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HeatMap.R
\name{brush_action_fullHM_LogFC}
\alias{brush_action_fullHM_LogFC}
\title{brush_action_fullHM_LogFC}
\usage{
brush_action_fullHM_LogFC(df, input, output, session)
}
\arguments{
\item{df}{data.frame containing data which is enclosed by brush action}

\item{input}{shiny input object}

\item{output}{shiny output object}

\item{session}{shiny session object}
}
\value{
nothing; function creates SPLOM
}
\description{
brush_action_fullHM_LogFC
}
