% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotHM.R
\name{brush_action_fullHM_P_Val}
\alias{brush_action_fullHM_P_Val}
\title{brush_action_fullHM_P_Val}
\usage{
brush_action_fullHM_P_Val(df, input, output, session)
}
\arguments{
\item{df}{data.frame containing data which is enclosed by brush action}

\item{input}{shiny input object}

\item{output}{shiny output object}

\item{session}{shiny session object}
}
\value{
nothing; function creates SPLOM
}
\description{
brush_action_fullHM_P_Val
}
