% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalculateModels.R
\name{calculateEpigeneticEffectsAdjusted}
\alias{calculateEpigeneticEffectsAdjusted}
\title{calculate multiple models in parallel}
\usage{
calculateEpigeneticEffectsAdjusted(ExposureVariable, config)
}
\description{
function to calculate multiple models in parallel for given ExposureVariables
}
