% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalculateModels.R
\name{CalculateModels}
\alias{CalculateModels}
\title{calculate models}
\usage{
CalculateModels(workDir)
}
\arguments{
\item{workDir}{working directory to look for config.yml in}
}
\description{
function to calculate models defined in the config.yml of the working directory outside the shiny interactive mode using multip
}
