% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ResultData.R
\name{checklistResultP_Val_DeltaMeth_NValidity}
\alias{checklistResultP_Val_DeltaMeth_NValidity}
\title{checklistResultP_Val_DeltaMeth_NValidity
checks for validity of structure}
\usage{
checklistResultP_Val_DeltaMeth_NValidity(listResultP_Val_DeltaMeth_N)
}
\arguments{
\item{listResultP_Val_DeltaMeth_N}{structure to check}
}
\value{
TRUE or FALSE
}
\description{
checklistResultP_Val_DeltaMeth_NValidity
checks for validity of structure
}
