% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ResultData.R
\name{checkResultP_Val_cg}
\alias{checkResultP_Val_cg}
\title{checkResultP_Val_cg
checks whether rownames of results start with 'cg'}
\usage{
checkResultP_Val_cg(listResultP_Val)
}
\arguments{
\item{listResultP_Val}{structure to check}
}
\value{
TRUE or FALSE
}
\description{
checkResultP_Val_cg
checks whether rownames of results start with 'cg'
}
