% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HeatMap.R
\name{combinedDFInteractiveHeatMap}
\alias{combinedDFInteractiveHeatMap}
\title{creates a regular heatmap}
\usage{
combinedDFInteractiveHeatMap(
  id,
  combinedDF_Labels,
  dendProbes = NA,
  dendTraits = NA,
  Distances = NA,
  selectedRowIndicesYellow = NA,
  selectedColIndices = NA,
  selectedRowIndicesOrange = NA,
  session = session
)
}
\arguments{
\item{combinedDF_Labels}{list of data.frame and labels generated from function mergeDFP_Val_Labels()}

\item{dendProbes}{dendrogram (without labels) for probes (rows), generated externally and providing information for clustering of heatmap}

\item{dendTraits}{dendrogram (without labels) result for traits (columns), generated externally and providing information for clustering of heatmap}

\item{selectedRowIndicesYellow}{indicies of HM rows to mark in yellow color}

\item{selectedColIndices}{indicies of HM cols to mark (not used so far)}

\item{selectedRowIndicesOrange}{indicies of HM rows to mark in orange color}

\item{session}{session object for reference}
}
\value{
heatmap object for InteractiveComplexHeatmap::makeInteractiveComplexHeatmap
}
\description{
creates a regular heatmap
}
