% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotHM.R
\name{combinedDFInteractiveHeatMapP_Val}
\alias{combinedDFInteractiveHeatMapP_Val}
\title{creates a regular heatmap}
\usage{
combinedDFInteractiveHeatMapP_Val(combinedDF_Labels, dendProbes, dendTraits)
}
\arguments{
\item{combinedDF_Labels}{list of data.frame and labels generated from function getCombinedDFP_Val_Labels()}

\item{dendProbes}{dendrogramProbes for probes (rows), generated externally and providing sorting information for heatmap#}

\item{dendTraits}{dendrogramTraits for traits (columns), generated externally and providing sorting information for heatmap}
}
\value{
heatmap object for InteractiveComplexHeatmap::makeInteractiveComplexHeatmap
}
\description{
creates a regular heatmap
}
