% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{extractMantissaExponent}
\alias{extractMantissaExponent}
\title{extractMantissaExponent}
\usage{
extractMantissaExponent(x)
}
\arguments{
\item{x}{value to extract mantissa and exponent}
}
\value{
list with mantissa and exponent from x
}
\description{
extractMantissaExponent
}
