% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{findInFiles}
\alias{findInFiles}
\title{findInFiles
finds strings in a bunch of files}
\usage{
findInFiles(
  what,
  where = ".",
  in_files = "\\\\.[Rr]$",
  recursive = TRUE,
  ignore.case = TRUE
)
}
\arguments{
\item{what}{what (string) to find}

\item{where}{directory, where to search}

\item{in_files}{regular expression for filenames, where to search}

\item{recursive}{recursive search}

\item{ignore.case}{ignore case sensitivity}
}
\value{
data frame with measured beta values
}
\description{
findInFiles
finds strings in a bunch of files
}
