% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ResultData.R
\name{getAvailNForP_VALBorder}
\alias{getAvailNForP_VALBorder}
\title{getAvailNForP_VALBorder
counts traits with at least 2 elements > P_VAL_BORDER}
\usage{
getAvailNForP_VALBorder(DF)
}
\arguments{
\item{DF}{data.frame with P_Val}
}
\value{
data.frame with reduced data set
}
\description{
getAvailNForP_VALBorder
counts traits with at least 2 elements > P_VAL_BORDER
}
