% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ResultData.R
\name{getAvailNForP_VALBorderParallel}
\alias{getAvailNForP_VALBorderParallel}
\title{getAvailNForP_VALBorderParallel
counts number of remaining models below a defined n; uses parallel processing for faster results}
\usage{
getAvailNForP_VALBorderParallel(session, wd, numCores, DF)
}
\arguments{
\item{numCores}{number of CPU cores to use}

\item{DF}{data frame for which to calculate results}
}
\value{
data.frame with results
}
\description{
getAvailNForP_VALBorderParallel
counts number of remaining models below a defined n; uses parallel processing for faster results
}
