% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SPLOM.R
\name{getBinaryFactorialVars}
\alias{getBinaryFactorialVars}
\title{getBinaryFactorialVars}
\usage{
getBinaryFactorialVars(df)
}
\arguments{
\item{df}{data frame}
}
\value{
list with variables that can be treated as factor in df
}
\description{
getBinaryFactorialVars
}
