% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Clustering.R
\name{getClustergramLong}
\alias{getClustergramLong}
\title{getClustergramLong
produces a clustergram}
\usage{
getClustergramLong(mat, clustRes, nClusters)
}
\arguments{
\item{mat}{transposed matrix with result values}

\item{clustRes}{(reduced) clustering results}

\item{nClusters}{number of cluster medoids}
}
\value{
plot with clustergram
examples getClustergramLong(mat, clustRes, nClusters)
}
\description{
getClustergramLong
produces a clustergram
}
