% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Clustering.R
\name{getClusterMedoids}
\alias{getClusterMedoids}
\title{getClusterMedoids
gets Medoids of earlier produced clusters; select trait with lowest (square) sum of distances as cluster medoid}
\usage{
getClusterMedoids(distMat, clustRes, numClusters)
}
\arguments{
\item{distMat}{distance matrix}

\item{clustRes}{(reduced) clustering results}

\item{numClusters}{number of clusters to produce}
}
\value{
list with cluster medoids
examples getClusterMedoids(distMat, clustRes, numClusters)
}
\description{
getClusterMedoids
gets Medoids of earlier produced clusters; select trait with lowest (square) sum of distances as cluster medoid
}
