% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Clustering.R
\name{getClustersTable}
\alias{getClustersTable}
\title{getClustersTable
gets a table earlier produced clusters}
\usage{
getClustersTable(id, listClusters, listMedoids)
}
\arguments{
\item{listClusters}{list with clusters}

\item{listMedoids}{list with medoids}
}
\value{
data.frame with information on clusters and its medoids
examples getClustersTable(listClusters, listMedoids)
}
\description{
getClustersTable
gets a table earlier produced clusters
}
