% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Clustering.R
\name{getClustResFast}
\alias{getClustResFast}
\title{getClustResFast
calculates hierarchical clustering using fastcluster}
\usage{
getClustResFast(distanceMatrix)
}
\arguments{
\item{distanceMatrix}{distance matrix}
}
\value{
hclust object
examples getClustResFast(distanceMatrix)
}
\description{
getClustResFast
calculates hierarchical clustering using fastcluster
}
