% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotHM.R
\name{getCombinedDFP_Val_Labels}
\alias{getCombinedDFP_Val_Labels}
\title{getCombinedDFP_Val_Labels
merges data.frames out of sessionVariable containing p-values, delta methylation values and N by probeID}
\usage{
getCombinedDFP_Val_Labels(sessionVariables, minN)
}
\arguments{
\item{sessionVariables}{sessionVariables}

\item{minN}{minimum n}
}
\value{
named list of data.frames, one df for merged P_Val, one for merged DeltaMethylation, one for merged N as well as labels:

result$dfP_Val for p-values

result$dfDM for delta methylation values

result$dfN for n

result$labelsDF1 for labels belonging to original df1

result$labelsDF2 for labels belonging to original df2

result$labelsDF3 for labels belonging to original df3
}
\description{
getCombinedDFP_Val_Labels
merges data.frames out of sessionVariable containing p-values, delta methylation values and N by probeID
}
