% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Clustering.R
\name{getDend}
\alias{getDend}
\title{getDend
gets a dendrogram with clustering results}
\usage{
getDend(clustRes, nClusters)
}
\arguments{
\item{clustRes}{clustering results for traits}

\item{nClusters}{list with traits for coloring}
}
\value{
dendrogram
examples getDendTraits(clustRes, nClusters)
}
\description{
getDend
gets a dendrogram with clustering results
}
