% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reduceTrait.R
\name{getDendTraits}
\alias{getDendTraits}
\title{getDendTraits
gets a dendrogram with clustering results of traits}
\usage{
getDendTraits(clustResTraits, traitClusters)
}
\arguments{
\item{clustResTraits}{clustering results for traits}

\item{traitClusters}{list with traits for coloring}
}
\value{
dendrogram
examples getDendTraits(clustResTraits, traitClusters)
}
\description{
getDendTraits
gets a dendrogram with clustering results of traits
}
