% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Clustering.R
\name{getDistMat}
\alias{getDistMat}
\title{getDistMat
calculates distance matrix using parallelDist}
\usage{
getDistMat(numberCores, matrix)
}
\arguments{
\item{numberCores}{number of CPU cores to use}

\item{matrix}{matrix to calculate distance matrix for}
}
\value{
distance matrix
examples getDistMat(numberCores, matrix)
}
\description{
getDistMat
calculates distance matrix using parallelDist
}
