% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotHM.R
\name{getlistOfResultsDF}
\alias{getlistOfResultsDF}
\title{getlistOfResultsDF
gets list of result data.frame from folder}
\usage{
getlistOfResultsDF(folder)
}
\arguments{
\item{folder}{folder containing all files to read to list}
}
\value{
list of data.frame from folder
}
\description{
getlistOfResultsDF
gets list of result data.frame from folder
}
