% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reduceTrait.R
\name{getMedoidsTable}
\alias{getMedoidsTable}
\title{getMedoidsTable
gets a table with Medoids of earlier produced clusters}
\usage{
getMedoidsTable(listMedoids)
}
\arguments{
\item{listMedoids}{list with medoids}
}
\value{
data.frame with information on medoids
examples getMedoidsTable(listMedoids)
}
\description{
getMedoidsTable
gets a table with Medoids of earlier produced clusters
}
