% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ResultData.R
\name{getNForDMBorder}
\alias{getNForDMBorder}
\title{getNForDMBorder}
\usage{
getNForDMBorder(mat, DMBorder)
}
\arguments{
\item{mat}{matrix for which to calculate results}

\item{n}{minimum n for which to calculate results}
}
\value{
data.frame with results
}
\description{
counts number of remaining models below a defined n based on delta methylation
}
