% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ResultData.R
\name{getNForNBorder}
\alias{getNForNBorder}
\title{getNForNBorder}
\usage{
getNForNBorder(mat, NBorder)
}
\arguments{
\item{mat}{matrix for which to calculate results}

\item{n}{minimum n for which to calculate results}
}
\value{
data.frame with results
}
\description{
counts number of remaining models below a defined n based on n
}
