% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ResultData.R
\name{getNForP_ValBorder}
\alias{getNForP_ValBorder}
\title{getNForP_ValBorder
counts number of remaining models below a defined n}
\usage{
getNForP_ValBorder(mat, n)
}
\arguments{
\item{mat}{matrix for which to calculate results}

\item{n}{minimum n for which to calculate results}
}
\value{
data.frame with results
}
\description{
getNForP_ValBorder
counts number of remaining models below a defined n
}
