% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ResultData.R
\name{getPHENODF}
\alias{getPHENODF}
\title{getPHENODF
loads a pheno DF for a certain PHENOFileName}
\usage{
getPHENODF(PHENOFileName, listPrimaryKeys)
}
\arguments{
\item{PHENOFileName}{filename for pheno DF to load}

\item{listPrimaryKeys}{list with possible primary keys in original data}
}
\value{
pheno DF for a certain PHENOFileName
}
\description{
getPHENODF
loads a pheno DF for a certain PHENOFileName
}
