% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Clustering.R
\name{getplotClustergramTraitsLong}
\alias{getplotClustergramTraitsLong}
\title{getplotClustergramTraitsLong
produces a clustergram for traits}
\usage{
getplotClustergramTraitsLong(mat.t, clustResTraits, traitClusters)
}
\arguments{
\item{mat.t}{transposed matrix with result values}

\item{clustResTraits}{(reduced) clustering results for traits}

\item{traitClusters}{trait cluster medoids}
}
\value{
plot with clustergram
examples getplotClustergramTraitsLong(mat.t, clustResTraits, traitClusters)
}
\description{
getplotClustergramTraitsLong
produces a clustergram for traits
}
