% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reducePValue.R
\name{getPReducedTraitData}
\alias{getPReducedTraitData}
\title{getPReducedTraitData
reduces data in structure combinedDFP_Val_Labels by p-value range and minimum n; entire CpGs will be removed}
\usage{
getPReducedTraitData(
  session,
  combinedDFP_Val_Labels,
  minP_Val,
  maxP_Val,
  minDM,
  maxDM,
  minN,
  maxN,
  debugMode
)
}
\arguments{
\item{combinedDFP_Val_Labels}{data structure which should become p_val-reduced}

\item{minP_Val}{minimum p-value to retain}

\item{maxP_Val}{maximum p-value to retain}

\item{minN}{minimum n for model in regression result}

\item{debugMode}{return smaller data structure (only session$userData$sessionVariables$debugNumber records) for faster debugging}
}
\value{
result list()
}
\description{
getPReducedTraitData
reduces data in structure combinedDFP_Val_Labels by p-value range and minimum n; entire CpGs will be removed
}
