% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Clustering.R
\name{getProbeReducedDataStructure}
\alias{getProbeReducedDataStructure}
\title{getProbeReducedDataStructure
delivers clustering results for probes as well as probes that are within a defined range (DNAdistances) around each selected probe}
\usage{
getProbeReducedDataStructure(id, input, output, session)
}
\arguments{
\item{id}{either "PVal" or "LogFC"}

\item{input}{shiny input object}

\item{output}{shiny output object}

\item{session}{shiny session object}
}
\value{
nested list with results of probe clustering
examples getProbeReducedDataStructure(id, input, output, session)
}
\description{
getProbeReducedDataStructure
delivers clustering results for probes as well as probes that are within a defined range (DNAdistances) around each selected probe
}
