% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ResultData.R
\name{getReducedP_Valdf}
\alias{getReducedP_Valdf}
\title{getReducedP_ValMatrix
counts and gets back traits with at least 2 elements > P_VAL_BORDER}
\usage{
getReducedP_Valdf(df, numRows, upperP_VALborder, lowerP_VALborder)
}
\arguments{
\item{df}{data.frame with P_Val, rows are probes, cols are traits}

\item{numRows}{scalar with maximum allowed row number}

\item{upperP_VALborder}{scalar with maximum allowed P_Val}

\item{lowerP_VALborder}{optionalscalar with minimum allowed P_Val}
}
\value{
matrix with reduced data set
}
\description{
getReducedP_ValMatrix
counts and gets back traits with at least 2 elements > P_VAL_BORDER
}
