% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotHM.R
\name{getResultDfP_D_N}
\alias{getResultDfP_D_N}
\title{getResultDfP_D_N}
\usage{
getResultDfP_D_N(listOfResultDF, P_D_N)
}
\arguments{
\item{listOfResultDF}{data.frame containing a list of data.frames to be merged}

\item{P_D_N}{scalar value "P", "D" or "N" describing whether to merge P_VAL, DeltaMeth or N}
}
\value{
merged data.frame
}
\description{
getResultDfP_D_N
}
