% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotHM.R
\name{getSelectedOriginalData}
\alias{getSelectedOriginalData}
\title{getSelectedOriginalData}
\usage{
getSelectedOriginalData(
  combinedDFP_Val_Labels,
  row_index,
  column_index,
  session
)
}
\arguments{
\item{combinedDFP_Val_Labels}{list with datastructure pointing to original data}

\item{row_index}{indicies of selection}

\item{column_index}{indicies of selection}

\item{session}{shiny session object}
}
\value{
df with merged original data
}
\description{
getSelectedOriginalData
}
