% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SPLOM.R
\name{getSPLOM}
\alias{getSPLOM}
\title{getSPLOM}
\usage{
getSPLOM(df, markingVar)
}
\arguments{
\item{df}{data frame, from which the SPLOM will be drawn}

\item{markingVar}{variable inside the data frame, which will be not drawn in SPLOM, instead it will be used for visual stratification}
}
\value{
a plotly figure, which can be drawn using renderPlotly
}
\description{
getSPLOM
}
