% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reduceTrait.R
\name{getTraitClusterMedoids}
\alias{getTraitClusterMedoids}
\title{getTraitClusterMedoids
gets Medoids of earlier produced clusters}
\usage{
getTraitClusterMedoids(clustResTraits, distMatTraits, numClusters)
}
\arguments{
\item{clustResTraits}{(reduced) clustering results of traits}

\item{distMatTraits}{(reduced) distance matrix of traits}

\item{numClusters}{number of clusters to produce}
}
\value{
list with cluster medoids
examples getTraitClusterMedoids(clustResTraits, distMatTraits, numClusters)
}
\description{
getTraitClusterMedoids
gets Medoids of earlier produced clusters
}
