% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reduceTrait.R
\name{getTraitReducedcombinedDFP_Val_Labels}
\alias{getTraitReducedcombinedDFP_Val_Labels}
\title{getTraitReducedcombinedDFP_Val_Labels
gets a list with data structure for regression results after reduction for traits}
\usage{
getTraitReducedcombinedDFP_Val_Labels(
  pReducedcombinedDFP_Val_Labels,
  ClusterMedoids,
  keys
)
}
\arguments{
\item{pReducedcombinedDFP_Val_Labels}{list with data structure for regression results (from p-val reduction)}

\item{ClusterMedoids}{medoids of clusters for name setting}

\item{keys}{key attributes for merging of original data}
}
\value{
list with data structure for regression model results
examples getTraitReducedcombinedDFP_Val_Labels(pReducedcombinedDFP_Val_Labels, ClusterMedoids, keys)
}
\description{
getTraitReducedcombinedDFP_Val_Labels
gets a list with data structure for regression results after reduction for traits
}
