% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reduceTrait.R
\name{getTraitReducedData}
\alias{getTraitReducedData}
\title{getTraitReducedData
gets a list with data structure for regression results after reduction for traits}
\usage{
getTraitReducedData(pReducedcombinedDFP_Val_Labels, ClusterMedoids, keys)
}
\arguments{
\item{pReducedcombinedDFP_Val_Labels}{list with data structure for regression results (from p-val reduction)}

\item{ClusterMedoids}{medoids of clusters for name setting}

\item{keys}{key attributes for merging of original data}
}
\value{
list with data structure for regression model results
examples getTraitReducedData(pReducedcombinedDFP_Val_Labels, ClusterMedoids, keys)
}
\description{
getTraitReducedData
gets a list with data structure for regression results after reduction for traits
}
