% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Clustering.R
\name{getTraitReducedDataStructure}
\alias{getTraitReducedDataStructure}
\title{getTraitReducedDataStructure
delivers clustering results for traits}
\usage{
getTraitReducedDataStructure(id, input, output, session)
}
\arguments{
\item{id}{either "PVal" or "LogFC"}

\item{input}{shiny input object}

\item{output}{shiny output object}

\item{session}{shiny session object}
}
\value{
nested list with results of trait clustering
examples getTraitReducedDataStructure(id, input, output, session)
}
\description{
getTraitReducedDataStructure
delivers clustering results for traits
}
