% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reduceTrait.R
\name{getTraitSubsetcombinedDFP_Val_Labels}
\alias{getTraitSubsetcombinedDFP_Val_Labels}
\title{getTraitSubsetcombinedDFP_Val_Labels
omits traits from combinedDFP_Val_Labels structure, traits to retain are defined in param traits}
\usage{
getTraitSubsetcombinedDFP_Val_Labels(combinedDFP_Val_Labels, traits, keys)
}
\arguments{
\item{combinedDFP_Val_Labels}{structure, where traits should be omitted}

\item{traits}{traits to retain}

\item{keys}{key attributes to retain in data}
}
\value{
combinedDFP_Val_Labels structure list()
examples loadResultDF(combinedDFP_Val_Labels, traits)
}
\description{
getTraitSubsetcombinedDFP_Val_Labels
omits traits from combinedDFP_Val_Labels structure, traits to retain are defined in param traits
}
