% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{getYoungestFile}
\alias{getYoungestFile}
\title{getYoungestFile
gets youngest (file date and time) file from a certain folder
in use for checking, whether cached .rds-files have to be updated}
\usage{
getYoungestFile(folder)
}
\arguments{
\item{folder}{folder to check in}
}
\value{
file filename of youngest file in folder
}
\description{
getYoungestFile
gets youngest (file date and time) file from a certain folder
in use for checking, whether cached .rds-files have to be updated
}
