% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GlobalSelection.R
\name{GlobalSelection_SERVER}
\alias{GlobalSelection_SERVER}
\title{getSelectedOriginalData}
\usage{
GlobalSelection_SERVER(id, session)
}
\arguments{
\item{session}{shiny session object}

\item{selectedOnly}{if TRUE onloy those traits are returned, that were previously selected by the user}

\item{combinedDFP_Val_Labels}{list with datastructure pointing to original data -> session$userData$sessionVariables$probeReducedDataStructure()$combinedDFP_Val_Labels}
}
\value{
df with merged original data
getSelectedOriginalDataTraits

df with merged original data
getSelectedOriginalDataProbes

df with merged original data
}
\description{
getSelectedOriginalData
}
