% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotHM.R
\name{hover_action_HM}
\alias{hover_action_HM}
\title{hover_action_HM}
\usage{
hover_action_HM(df, input, output, session)
}
\arguments{
\item{df}{data.frame containing data which is enclosed by brush action}

\item{input}{shiny input object}

\item{output}{shiny output object}

\item{session}{shiny session object}
}
\value{
nothing; function info label for click action in HM
}
\description{
hover_action_HM
}
