% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{is.valid}
\alias{is.valid}
\title{is.valid
checks, whether an R object (x) is valid (not NULL) and exists}
\usage{
is.valid(x)
}
\arguments{
\item{x}{object}
}
\value{
TRUE or FALSE
examples is.valid(x)
}
\description{
is.valid
checks, whether an R object (x) is valid (not NULL) and exists
}
