% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CombinedData.R
\name{loadDir}
\alias{loadDir}
\title{loadDir}
\usage{
loadDir(session, traitDirList)
}
\arguments{
\item{session}{session info}

\item{traitDirList}{list containing directories to load into PatternMatchR}
}
\value{
data.frame with contents of traitDirList
}
\description{
loadDir
}
