% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{loadObjects}
\alias{loadObjects}
\title{loadObjects
loads globally needed objects (methylation matrix with beta values, annotation)}
\usage{
loadObjects(session)
}
\arguments{
\item{session}{session object}
}
\description{
loadObjects
loads globally needed objects (methylation matrix with beta values, annotation)
}
