% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalculateModels.R
\name{loadObjectsForCalculation}
\alias{loadObjectsForCalculation}
\title{load required data for model calculation}
\usage{
loadObjectsForCalculation(config)
}
\description{
function to load all data to calculate regression model into memory
}
