% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotHM.R
\name{loadResultDF}
\alias{loadResultDF}
\title{loadResultDF
loads data.frame from folder}
\usage{
loadResultDF(folder)
}
\arguments{
\item{folder}{folder containing all files to read to list}
}
\value{
list with resultP_Val, resultDeltaMethylation, resultN from folder
}
\description{
loadResultDF
loads data.frame from folder
}
