% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{loadtraitDFs}
\alias{loadtraitDFs}
\title{loadtraitDFs
loads corresponding data.frames for p_val (DFP_Val), delta methylation (DFDM) and n (DFN)}
\usage{
loadtraitDFs(traitDFs)
}
\arguments{
\item{traitDFs}{data.frames}
}
\value{
list with three data.frames
}
\description{
loadtraitDFs
loads corresponding data.frames for p_val (DFP_Val), delta methylation (DFDM) and n (DFN)
}
