% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HeatMap.R
\name{plotCombinedCondDWHM_P_Val}
\alias{plotCombinedCondDWHM_P_Val}
\title{plotCombinedCondDWHM_P_Val
plots condensed distance weighted heatmap}
\usage{
plotCombinedCondDWHM_P_Val(input, output, session)
}
\arguments{
\item{input}{shiny input}

\item{output}{shiny output}

\item{session}{shiny session}
}
\value{
nothing
examples plotCombinedCondDWHM_P_Val(input, output, session)
}
\description{
plotCombinedCondDWHM_P_Val
plots condensed distance weighted heatmap
}
