% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotHM.R
\name{plotCombinedDWHM_P_Val}
\alias{plotCombinedDWHM_P_Val}
\title{plotCombinedDWHM_P_Val
plots distance weighted heatmap}
\usage{
plotCombinedDWHM_P_Val(input, output, session)
}
\arguments{
\item{input}{shiny input}

\item{output}{shiny output}

\item{session}{shiny session}
}
\value{
nothing
examples plotCombinedDWHM_P_Val(input, output, session)
}
\description{
plotCombinedDWHM_P_Val
plots distance weighted heatmap
}
