% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{removeAdjFromColname}
\alias{removeAdjFromColname}
\title{removeAdjFromColname}
\usage{
removeAdjFromColname(colnames)
}
\arguments{
\item{colnames}{list with colnames probably containing 'adj' at the end of each name}
}
\value{
colnames without 'adj' at the end
adj' was added to the end of the name of an analysis for a regression model with additional adjustment variables
}
\description{
removeAdjFromColname
}
