% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ResultData.R
\name{removeTraitsMinN}
\alias{removeTraitsMinN}
\title{removeTraitsMinN
removes CpG with casecount < minN from corresponding data.frames for P_Val, for DeltaMethylation and for N}
\usage{
removeTraitsMinN(dfList, minN)
}
\arguments{
\item{dfList}{list containing corresponding data.frames for P_Val, for DeltaMethylation and for N}

\item{minN}{minimum value for n}
}
\value{
named list of data.frames, one df for P_Val, one for DeltaMethylation, one for N as well as labels

result$dfP_Val for p-values

result$dfDM for delta methylation values

result$dfN for n
}
\description{
removeTraitsMinN
removes CpG with casecount < minN from corresponding data.frames for P_Val, for DeltaMethylation and for N
}
