% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Clustering.R
\name{updateTxtClusterOut}
\alias{updateTxtClusterOut}
\title{updateTxtClusterOut
generates summary text after clustering}
\usage{
updateTxtClusterOut(
  traitReducedcombinedDFP_Val_Labels,
  minP_Val,
  maxP_Val,
  minN,
  sldNumClasses
)
}
\arguments{
\item{traitReducedcombinedDFP_Val_Labels}{data structure with trait reduced results}

\item{minP_Val}{minimum p_value for model to use for clustering}

\item{maxP_Val}{maximum p_value for model to use for clustering}

\item{minN}{minimum n for model to use for clustering}

\item{sldNumClasses}{number of classes to use for clustering}
}
\value{
text
examples updateTxtClusterOut(traitReducedcombinedDFP_Val_Labels, minP_Val, maxP_Val, minN, sldNumClasses)
}
\description{
updateTxtClusterOut
generates summary text after clustering
}
