% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CombinedData.R
\name{updateTxtLoadOut}
\alias{updateTxtLoadOut}
\title{updateTxtLoadOut}
\usage{
updateTxtLoadOut(
  session,
  resultDFListTrait1,
  resultDFListTrait2,
  resultDFListTrait3
)
}
\arguments{
\item{resultDFListTrait1}{data.frame containing results from trait1 (red)}

\item{resultDFListTrait2}{data.frame containing results from trait2 (green)}

\item{resultDFListTrait3}{data.frame containing results from trait3 (blue)}
}
\value{
HTML to show in result info line after data load section of PatternMatchR
}
\description{
updateTxtLoadOut
}
