% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CombinedData.R
\name{updateTxtMergeOut}
\alias{updateTxtMergeOut}
\title{updateTxtMergeOut}
\usage{
updateTxtMergeOut(combinedDFP_Val_Labels)
}
\arguments{
\item{combinedDFP_Val_Labels}{data.frame containing merged results}
}
\value{
HTML to show in result info line after data merge section of PatternMatchR
}
\description{
updateTxtMergeOut
}
