% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reducePValue.R
\name{updateTxtpReduceOut}
\alias{updateTxtpReduceOut}
\title{updateTxtpReduceOut
generates summary text after p-value reduction}
\usage{
updateTxtpReduceOut(pReducedcombinedDFP_Val_Labels)
}
\arguments{
\item{pReducedcombinedDFP_Val_Labels}{data structure with trait reduced results}
}
\value{
text
examples updateTxtpReduceOut(pReducedcombinedDFP_Val_Labels)
}
\description{
updateTxtpReduceOut
generates summary text after p-value reduction
}
