% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotHM.R
\name{combinedDFInteractiveHeatMapP_Val}
\alias{combinedDFInteractiveHeatMapP_Val}
\title{creates a regular heatmap}
\usage{
combinedDFInteractiveHeatMapP_Val(
  combinedDF_Labels,
  dendProbes = NA,
  dendTraits = NA,
  selectedRowIndices = NA,
  selectedColIndices = NA
)
}
\arguments{
\item{combinedDF_Labels}{list of data.frame and labels generated from function mergeDFP_Val_Labels()}

\item{dendProbes}{dendrogram (without labels) for probes (rows), generated externally and providing information for clustering of heatmap}

\item{dendTraits}{dendrogram (without labels) result for traits (columns), generated externally and providing information for clustering of heatmap}
}
\value{
heatmap object for InteractiveComplexHeatmap::makeInteractiveComplexHeatmap
}
\description{
creates a regular heatmap
}
