% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ResultData.R
\name{delete.na}
\alias{delete.na}
\title{delete.na
deletes rows with all NA from data.frame DF}
\usage{
delete.na(df, n = 0)
}
\arguments{
\item{df}{data.frame}

\item{n}{minimum number of NA}
}
\value{
data.frame
}
\description{
delete.na
deletes rows with all NA from data.frame DF
}
