% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SPLOM.R
\name{getDimensionsForPlotlySPLOM}
\alias{getDimensionsForPlotlySPLOM}
\title{structure of sessionVariables for SPLOM:
detailed original data for SPLOM analysis
session$userData$sessionVariables$fullSPLOM - data structure for scatter plot matrix
session$userData$sessionVariables$selectedFullOriginalData - merged original data for creation of SPLOM
session$userData$sessionVariables$fullMarkingVar - variable, which should be marked (stratified for) in SPLOM
getDimensionsForPlotlySPLOM}
\usage{
getDimensionsForPlotlySPLOM(df, omitVar)
}
\arguments{
\item{df}{data frame}

\item{omitVar}{Variable to leave out of nested list}
}
\value{
nested list for plotly with variable labels and formulae
}
\description{
structure of sessionVariables for SPLOM:
detailed original data for SPLOM analysis
session$userData$sessionVariables$fullSPLOM - data structure for scatter plot matrix
session$userData$sessionVariables$selectedFullOriginalData - merged original data for creation of SPLOM
session$userData$sessionVariables$fullMarkingVar - variable, which should be marked (stratified for) in SPLOM
getDimensionsForPlotlySPLOM
}
