% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reduceTrait.R
\name{getplotClustergramTraitsLong}
\alias{getplotClustergramTraitsLong}
\title{getplotClustergramTraitsLong
produces a clustergram for traits}
\usage{
getplotClustergramTraitsLong(matP_Val.t, clustResTraits, traitClusters)
}
\arguments{
\item{matP_Val.t}{transposed matrix with result p-values}

\item{clustResTraits}{(reduced) clustering results for traits}

\item{traitClusters}{trait cluster medoids}
}
\value{
plot with clustergram
examples getplotClustergramTraitsLong(matP_Val.t, clustResTraits, traitClusters)
}
\description{
getplotClustergramTraitsLong
produces a clustergram for traits
}
