% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CombinedData.R
\name{getResultDfP_D_N}
\alias{getResultDfP_D_N}
\title{combinedDFP_Val_Labels$dfP_Val - data frame with p-values from regression results
combinedDFP_Val_Labels$dfDM - data frame with delta methylations from regression results
combinedDFP_Val_Labels$dfN - data frame with n from regression results
combinedDFP_Val_Labels$labelsDF1 - variable labels from data frame for trait 1
combinedDFP_Val_Labels$labelsDF2 - variable labels from data frame for trait 2
combinedDFP_Val_Labels$labelsDF3 - variable labels from data frame for trait 3
combinedDFP_Val_Labels$labelsDF <- c(LabelsDF1, LabelsDF2, LabelsDF3) - combined variable labels for all three traits
combinedDFP_Val_Labels$mergedOriginDF - number (label) of original data frame
combinedDFP_Val_Labels$mergedColnames <- merged original Colnames
combinedDFP_Val_Labels$mergedOriginTrait <- number of trait (1,2 or 3), a particular variable belongs to
getResultDfP_D_N}
\usage{
getResultDfP_D_N(listOfResultDF, P_D_N)
}
\arguments{
\item{listOfResultDF}{data.frame containing a list of data.frames to be merged}

\item{P_D_N}{scalar value "P", "D" or "N" describing whether to merge P_VAL, DeltaMeth or N}
}
\value{
merged data.frame
}
\description{
combinedDFP_Val_Labels$dfP_Val - data frame with p-values from regression results
combinedDFP_Val_Labels$dfDM - data frame with delta methylations from regression results
combinedDFP_Val_Labels$dfN - data frame with n from regression results
combinedDFP_Val_Labels$labelsDF1 - variable labels from data frame for trait 1
combinedDFP_Val_Labels$labelsDF2 - variable labels from data frame for trait 2
combinedDFP_Val_Labels$labelsDF3 - variable labels from data frame for trait 3
combinedDFP_Val_Labels$labelsDF <- c(LabelsDF1, LabelsDF2, LabelsDF3) - combined variable labels for all three traits
combinedDFP_Val_Labels$mergedOriginDF - number (label) of original data frame
combinedDFP_Val_Labels$mergedColnames <- merged original Colnames
combinedDFP_Val_Labels$mergedOriginTrait <- number of trait (1,2 or 3), a particular variable belongs to
getResultDfP_D_N
}
