% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotHM.R
\name{getSelectedOriginalData}
\alias{getSelectedOriginalData}
\title{structure of sessionVariables:
result data for/from clustering
session$userData$sessionVariables$distMatProbes - distance matrix for probes out of p-value results of regression
session$userData$sessionVariables$clustResProbes - clustering reusult for probes out of p-value results of regression
session$userData$sessionVariables$distMatTraits - distance matrix for traits out of p-value results of regression
session$userData$sessionVariables$clustResTraits - clustering reusult for traits out of p-value results of regression
session$userData$sessionVariables$combinedHMP_Val - data structure for combined heatmap
getSelectedOriginalData}
\usage{
getSelectedOriginalData(
  combinedDFP_Val_Labels,
  row_index,
  column_index,
  session
)
}
\arguments{
\item{combinedDFP_Val_Labels}{list with datastructure pointing to original data}

\item{row_index}{indicies of selection}

\item{column_index}{indicies of selection}

\item{session}{shiny session object}
}
\value{
df with merged original data
}
\description{
structure of sessionVariables:
result data for/from clustering
session$userData$sessionVariables$distMatProbes - distance matrix for probes out of p-value results of regression
session$userData$sessionVariables$clustResProbes - clustering reusult for probes out of p-value results of regression
session$userData$sessionVariables$distMatTraits - distance matrix for traits out of p-value results of regression
session$userData$sessionVariables$clustResTraits - clustering reusult for traits out of p-value results of regression
session$userData$sessionVariables$combinedHMP_Val - data structure for combined heatmap
getSelectedOriginalData
}
