% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{loadDNAm}
\alias{loadDNAm}
\title{loadDNAm
loads measured beta values from file config$betaFileName which was defined in configuration}
\usage{
loadDNAm(config)
}
\arguments{
\item{config}{configuration}
}
\value{
data frame with measured beta values
}
\description{
loadDNAm
loads measured beta values from file config$betaFileName which was defined in configuration
}
