% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ResultData.R
\name{loadFolderDFList}
\alias{loadFolderDFList}
\title{loadFolderDFList
loads list of data.frame from folder}
\usage{
loadFolderDFList(session, folder)
}
\arguments{
\item{session}{session object}

\item{folder}{folder containing all files to read to list}
}
\value{
list of data.frame from folder
}
\description{
loadFolderDFList
loads list of data.frame from folder
}
