% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{loadResultFile}
\alias{loadResultFile}
\title{loadResultFile}
\usage{
loadResultFile(session, folder, fileName)
}
\arguments{
\item{session}{session object}

\item{folder}{folder containing file with fileName}

\item{fileName}{file containing probeID, p-values, delta methylation-values and n from regression}
}
\value{
data.frame with regression results from file
}
\description{
loadResultFile
}
