% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ResultData.R
\name{loadtraitDFs}
\alias{loadtraitDFs}
\title{loadtraitDFs
loads and concatenates corresponding data.frames for p_val (DFP_Val), delta methylation (DFDM) and n (DFN)}
\usage{
loadtraitDFs(traitDFs)
}
\arguments{
\item{traitDFs}{data.frames}
}
\value{
result list()
data.frame P_Val,
data.frame DM (DeltaMethylation),
data.frame N,
data.frame PHENOdata,
chr PHENOFileName,
list resultOriginDF
list resultColnames
}
\description{
loadtraitDFs
loads and concatenates corresponding data.frames for p_val (DFP_Val), delta methylation (DFDM) and n (DFN)
}
