% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CombinedData.R
\name{mergeDFP_Val_Labels}
\alias{mergeDFP_Val_Labels}
\title{mergeDFP_Val_Labels
merges data.frames out of sessionVariable containing p-values, delta methylation values and N by probeID}
\usage{
mergeDFP_Val_Labels(
  resultDFListTrait1,
  resultDFListTrait2,
  resultDFListTrait3,
  minN
)
}
\arguments{
\item{resultDFListTrait1}{trait structure1 to merge}

\item{resultDFListTrait2}{trait structure2 to merge}

\item{resultDFListTrait3}{trait structure3 to merge}

\item{minN}{minimum n}
}
\value{
named list of data.frames, one df for merged P_Val, one for merged DeltaMethylation, one for merged N as well as labels:

result$dfP_Val for p-values

result$dfDM for delta methylation values

result$dfN for n
}
\description{
mergeDFP_Val_Labels
merges data.frames out of sessionVariable containing p-values, delta methylation values and N by probeID
}
