% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HeatMap.R
\name{plotCombinedCondHM_P_Val}
\alias{plotCombinedCondHM_P_Val}
\title{plotCombinedCondHM_P_Val
plots condensed heatmap}
\usage{
plotCombinedCondHM_P_Val(input, output, session)
}
\arguments{
\item{input}{shiny input}

\item{output}{shiny output}

\item{session}{shiny session}
}
\value{
nothing
examples plotCombinedCondHM_P_Val(input, output, session)
}
\description{
plotCombinedCondHM_P_Val
plots condensed heatmap
}
